Remove duplicates, rename, create link
I didn't actually remove anything, just renamed old/duplicate
simulateAE and simulateDst scripts, and left behind the ones
that were pointing to deltaBTimeSeriespara.py, since those
seemed to be the ones used most recently.

I then "deleted" a duplicate of deltaBTimeSeries.py in the
scripts folder, and created a symlink to the working copy in
the Tools folder. The latter should be functionally equivalent
to the former.
