+++ pyLTR/Tools/simulateDst_old_deleteme.py
